requirements.txt
absl-py==1.4.0
africastalking==1.2.5
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
ansi2html==1.8.0
APScheduler==3.10.4
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asn1crypto==1.5.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
autopep8==1.4.4
Babel==2.12.1
base58check==1.0.2
bash==0.6
beautifulsoup4==4.12.2
billiard==3.6.4.0
bitcoin==1.1.42
blinker==1.6.2
Brotli==1.1.0
bs4==0.0.1
cachelib==0.9.0
cachetools==5.3.1
cancelchain==1.4.0
celery==5.2.7
certifi==2024.2.2
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.1.1
cron-descriptor==1.4.0
crypto==1.4.1
cryptography==41.0.1
cryptomus==1.1
cssselect2==0.7.0
cycler==0.11.0
dash==2.13.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
dateparser==1.1.8
defusedxml==0.6.0
distlib==0.3.7
dj-database-url==2.1.0
Django==5.0.4
django-admin-black==1.0.8
django-allauth==0.54.0
django-browser-reload==1.13.0
django-celery-beat==2.5.0
django-core==1.4.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-daraja==1.3.0
django-heroku==0.3.1
django-payments==2.0.0
django-phonenumber-field==7.1.0
django-restframework==0.0.1
django-timezone-field==5.1
django-widget-tweaks==1.4.8
dnspython==2.4.2
docutils==0.20.1
ecdsa==0.18.0
email-validator==2.0.0.post2
filelock==3.12.2
Flask==2.2.5
Flask-BabelEx==0.9.4
Flask-Caching==2.0.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
flatbuffers==23.5.26
fonttools==4.42.1
forex-python==1.8
frozenlist==1.3.3
gast==0.4.0
gevent==23.9.1
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-pasta==0.2.0
greenlet==3.0.0
grpcio==1.58.0
gunicorn==21.2.0
h5py==3.9.0
html5lib==1.1
humanfriendly==10.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.13.1
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
kombu==5.2.4
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
mathpy==0.3.0
matplotlib==3.8.0
mdurl==0.1.2
millify==0.1.1
mpesa-connect==0.1.4
multidict==6.0.4
mysql==0.0.3
mysqlclient==2.1.1
Naked==0.1.32
nest-asyncio==1.5.8
npm==0.1.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.7
opt-einsum==3.3.0
optional-django==0.1.0
oscrypto==1.3.0
packaging==24.0
pandas==2.1.1
passlib==1.7.4
pep8==1.7.1
pg8000==1.29.4
phonenumberslite==8.13.18
pillow==10.3.0
pipenv==2023.7.23
platformdirs==3.10.0
plotly==5.17.0
prompt-toolkit==3.0.38
protobuf==4.24.3
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.5.0
pycparser==2.21
pycryptodome==3.17
pygame==2.5.0
Pygments==2.15.1
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.6.0
pymerkle==4.0.0
PyMySQL==1.0.3
pyparsing==3.1.1
pypdf==3.16.4
pypiwin32==223
pypng==0.20220715.0
pyreadline3==3.4.1
python-bidi==0.4.2
python-binance==1.0.19
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3
pywebio==1.8.2
pywin32==306
PyYAML==6.0
qrcode==7.4.2
regex==2023.8.8
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.3.5
rsa==4.9
schema==0.7.5
scramp==1.4.4
segno==1.5.2
shellescape==3.8.1
simplejson==3.19.1
six==1.16.0
sklearn==0.0.post7
soupsieve==2.4.1
speaklater==1.3
SQLAlchemy==1.4.48
sqlparse==0.5.0
str2bool==1.1
stripe==5.5.0
sudo==1.0.0
svglib==1.5.1
telegram==0.0.1
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tinycss2==1.2.1
tornado==6.3.2
tqdm==4.65.0
twilio==9.2.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
ua-parser==0.16.1
ujson==5.8.0
uritools==4.0.2
urllib3==2.2.1
user-agents==2.2.0
vine==5.0.0
virtualenv==20.24.2
virtualenv-clone==0.5.7
wcwidth==0.2.6
webencodings==0.5.1
websocket==0.2.1
websockets==11.0.3
Werkzeug==2.2.3
whatsapp-api-client-python==0.0.41
whitenoise==6.6.0
wrapt==1.15.0
WTForms==3.0.1
xhtml2pdf==0.2.11
yarl==1.9.2
zope.event==5.0
zope.interface==6.1
